// src/shared/enum/trend.enum.ts
export enum TrendEnum {
  UP = 'up',
  DOWN = 'down',
  SIDEWAY_EXPANDING = 'expanding sideway',
  SIDEWAY_CONTRACTING = 'contracting sideway',
  FLAT_SIDEWAY = 'flat sideway',
  WAIT_BREAKOUT = 'wait breakout',
  WAIT_BREAKDOWN = 'wait breakdown',
  MAYBE_UP = 'maybe up',
  MAYBE_DOWN = 'maybe down',
  UNSPECIFIED = 'unspecified',
}